requirements.txt
altair==4.1.0
blinker==1.4
dlib==19.22.1
face-recognition==1.3.0
gitdb==4.0.5
GitPython==3.1.11
html5lib==1.0.1
imbalanced-learn==0.7.0
ipaddr==2.2.0
ipykernel==5.3.4
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
matplotlib==3.2.2
nbclient==0.5.1
notebook==6.1.5
numpy==1.19.4
pandas==1.1.4
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.0.1
pip==21.3.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
pyzmq==20.0.0
six==1.15.0
scikit-learn==0.23.2
streamlit==0.71.0
urllib3==1.26.2
validators==0.18.1
watchdog==0.10.4
webencodings==0.5.1
widgetsnbextension==3.5.1
numpy>=1.14.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python>=4.5.5.64
opencv-contrib-python>=4.3.0.36
tensorflow>=1.9.0
keras>=2.2.0
Flask>=1.1.2
mtcnn>=0.1.0
lightgbm>=2.3.1
dlib>=19.20.0
retina-face>=0.0.1
mediapipe>=0.8.7.3
fire>=0.4.0
deepface==0.0.75
